Converting git repository into DNANexus App 	renamed:    resources/idConfigFieldRelationship.json -> eggd_qc_classifier/resources/home/dnanexus/app_resources/idConfigFieldRelationship.json 	renamed:    bin/utils.py -> eggd_qc_classifier/resources/home/dnanexus/bin/utils.py 	renamed:    full_assembly.py -> eggd_qc_classifier/resources/home/dnanexus/full_assembly.py     new folder:   eggd_qc_classifier/resources/home/dnanexus/packages/
