+++ eggd_qc_classifier/dxapp.json
@@ -1,0 +1,71 @@
+{
+  "name": "eggd_qc_classifier",
+  "title": "eggd_QC_classifier",
+  "summary": "An app to summarise important findings from MultiQC .json reports.",
+  "dxapi": "1.0.0",
+  "version": "0.0.1",
+  "inputSpec": [
+    {
+      "name": "sample_sheet",
+      "label": "SampleSheet.csv file of a run",
+      "class": "file",
+      "optional": false,
+      "patterns": [
+        "*.csv"
+      ],
+      "help": ""
+    },
+    {
+      "name": "multiqc_data",
+      "label": "Multiqc.json file of the associated run",
+      "class": "file",
+      "optional": false,
+      "patterns": [
+        "*.json"
+      ],
+      "help": ""
+    },
+    {
+      "name": "config_file",
+      "label": "config.yaml file of the type of sequencing run, or the yaml file with parameters of interest",
+      "class": "file",
+      "optional": false,
+      "patterns": [
+        "*.yaml"
+      ],
+      "help": ""
+    }
+  ],
+  "outputSpec": [
+    {
+      "name": "qc_report_json",
+      "label": "Json file with summarised metrics data",
+      "class": "file",
+      "patterns": [
+        "*"
+      ],
+      "help": ""
+    }
+  ],
+  "runSpec": {
+    "timeoutPolicy": {
+      "*": {
+        "hours": 1
+      }
+    },
+    "interpreter": "bash",
+    "file": "src/code.sh",
+    "distribution": "Ubuntu",
+    "release": "20.04",
+    "version": "0"
+  },
+  "regionalOptions": {
+    "aws:eu-central-1": {
+      "systemRequirements": {
+        "*": {
+          "instanceType": "mem1_ssd1_v2_x4"
+        }
+      }
+    }
+  }
+}
